v0.9.1 bug fixes
Apparently there was no circ2mat for QIRseq, that's fixed.  qcc
now works in a more top-down manner (see release notes).  UZERO,
the table of init values used by skalgo, now introduces the "I"
gate when the identity approximates a gate. This occurs when
global shift operators are used (all of which have SU(2)
representation as I).  Eventually, it'd be good to ignore/remove
any occurrence of I.