Added the QIR gate factory function QIR.
The function QIR can be used to construct QIR gates.  This
single point of access to QIR construction is the first step
in circuit construction without parsing descriptors by means of
horzcat (i.e. [a,b]) methods on QIR gates.  QIR still needs to
be tested for error checking.  All the usages of QIR need to be
documented in the usage text as well.