Added ctranspose methods to circuit classes.
In addition to a few bug fixes.  The ctranspose operator
was overloaded for all circuit classes. The effect of this
operator is to invert the circuit/gate, i.e. reverse order
and inverse operators. The method needs unit tests at all levels.
Some ad hoc testing was done at the REPL.