Changes to QIR and horzcat for seq. construction
A no-arg call to QIR no produces an empty circuit. The horzcat
method for QIR object now handles concatenation with empty circuits.
These two items combine allow for the construction of singleton
circuits with [QIR,a] or [a,QIR]. Unit tests are still needed, but
some crude testing at the REPL was carried out.

Changes were also made to horzcat so that the only case that
produces nested sequences is circuit-circuit concatenation.
Gate-circuit (and circuit-gate) concatenation will add the
gate to the top level of the circuit's sequence. Testing
 is still needed for this functionality.