QIR construction speed up
The QIRcircuit constructor is now a bit faster. It made
two calls to collectTars when it only needed one. CollectTars
for seq also was sped up quite a bit by removing repeated calls
to union and doing a single call to unique after collecting
all the targets in the sequence.