v0.9.8 parallelized table search in skalgo
The bottleneck of skalgo is the search through
UZERO for initial approximations. I tried different
serial implementations but the "loop and collect the
min as you go" version seemed to be optimal. Thus,
I tried parcellfun from the parallel package. This
led to some performance gains.  It will currently use
half the available cores.  Further work might be done
to find/compute the optimal chunks per proc for the
parcellfun launch. The current value of 5 was choosen
via trial and error and data collected by proffindclose
found in the skalgodata folder.