Performance improvements to SK compiler
findclosest is not set up for optional usage of the
parallel package.  The dependency between qcs and parallel
will probably be removed and the usage of parallel will be
encouraged. I saw as much as 2x speedup running findclosest
on 4 cores.  The UZERO table now has a third column which
contains the index of the conjugate for the row.  So
UZERO{k,3} is the index of the conjugate for the sequence given
in row k.  All rows are guaranteed to have their conjugate
in the table. Findcloset now returns the index for the approximating
sequence. skalgo returns a vector of indexes.  QIASMsingle/compile
then converts these vectors to sequences. This change got rid of
the computation of conjugate sequences in skalgo. This should result
in some speedup both from reduced computation and memory management.