UZERO with verified conjugates and pointers
addinverse transforms the 2 column UZERO to 3 column
uzero where the third column is the index of the conjugate.
It also removes any rows that do not have a conjugate in
the table.  Verifytable was updated to check that the
sequence computes the matrix (after global phase is factored out)
and that the conjugate row is, in fact a conjugate.