Bug fixes and Performance improve. for SK compiler
Continued work on QIASMsingle/compile, the SK compiler.  The
table in uzero.mat is now setup with a proper set of initial
sequences and their conjugates.  skalgo wasn't computing the
conjugate of group commutator sequences; that's fixed.  Some
more parallelism was added to the creation of the QASMseq
that results from the compiler.  This only results in a minor
speedup.  It seems that the current compiler works some of the
time.  Not all gates will meet approximation requirements, but
it seems the majority will.  It's still fairly slow, but it seems
like you can get a recursion depth 4 or 5 approximation in about
one minute (on my laptop using parallel).

In skalgodata several files were added that can be used in
analyzing and testing SK compiler related functions.