QIASMsingle/compile analysis
The SK compiler used by QIASMsingle has several
functions for testing, profiling, and algorithm
analysis in skalgodata.  It seems that on the
average the algorithm behaves admirably down to
recursion depth 4. At depth 5 the approximation
gets much much worse.  So, the current parameters are
set up around the idea that circuit error <= .009
is the target and the recursion depth should bottom
out at 4 unless you're compiling a number of gates
that you're just not going to sit around for.

The speed is still pretty slow. Decent speeds up
are obtained by loading the parallel package, but
we're still looking at something like a minute a gate
when run on 4/8 cores of my i7.  It will have to do for now.