Added new measure function. Resolves #20
There is now a measure function available to users. This function
performs a complete measurement in the standard basis on either
a state vector or a density matrix. The size of the space is
determined by the state representation.  By default, a single
measurement is performed and the output is given as an integer.
Users can optionally request binary output and multiple measurement
samples. When multiple samples are taken, then the majority result
is returned. This resolves issue #20.