Updated tarball for 1.2.2
I'm calling this good for help string documentation for the time
being.  The @QIR*, @QIASM*, and @QASM* classes all need documentation.
There's an incredible amount of redundancy in all these classes
due the hierarchy of the languages, so I'd like to find a way
to leverage this and not have to copy paste so much. Either way, I'd
rather use time on a written guide that documents what aspects of
these classes users are most likely to use directly and how
they'll use them.