new Rk constructor and updated QFT
There's now a QIR single constructor for the R(k) phase shift
used in the QFT and associated algorithms. For k<4, the QASM
equivalents are used (I,Z,S,T). Otherwise, the operator is computed
as a Phase/Amplitude operator.
The QFT function was updated to make use of this new constructor.